Applied suggestions, major result handling rewrite

+++ YellowNotes.Core/Services/NoteService.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -30,11 +31,21 @@ namespace YellowNotes.Core.Services
             return mapper.Map<NoteDto>(result);
         }
 
-        public async Task<object> GetNote(int noteId, string email,
+        public async Task<ResultHandler> GetNote(int noteId, string email,
             CancellationToken cancellationToken)
         {
             var note = await repository.GetNote(noteId, email, cancellationToken);
-            return mapper.Map<NoteDto>(note);
+
+            if (note == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (note.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+            return new ResultHandler(HttpStatusCode.OK, mapper.Map<NoteDto>(note));
         }
 
         public async Task<NotesDto> GetNotes(GetNotesConfig config, string email,
@@ -48,16 +59,48 @@ namespace YellowNotes.Core.Services
             };
         }
 
-        public async Task<object> UpdateNote(NoteDto note, string email,
+        public async Task<ResultHandler> UpdateNote(NoteDto note, string email,
             CancellationToken cancellationToken)
         {
-            return await repository.UpdateNote(note, email, cancellationToken);
+            var record = await repository.GetNote(note.NoteId, email, cancellationToken);
+
+            if (record == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (record.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+
+            var success = await repository.UpdateNote(note, email, cancellationToken);
+
+            return success
+                ? new ResultHandler(HttpStatusCode.NoContent)
+                : new ResultHandler(HttpStatusCode.UnprocessableEntity, "Failed to update note");
         }
 
-        public async Task<object> DeleteNote(int noteId, string email,
+        public async Task<ResultHandler> DeleteNote(int noteId, string email,
             CancellationToken cancellationToken)
         {
-            return await repository.DeleteNote(noteId, email, cancellationToken);
+            var note = await repository.GetNote(noteId, email, cancellationToken);
+
+            if (note == null)
+            {
+                return new ResultHandler(HttpStatusCode.NotFound);
+            }
+            else if (note.User.Email != email)
+            {
+                return new ResultHandler(HttpStatusCode.Unauthorized,
+                    "Requested resource is not available!");
+            }
+
+            var success = await repository.DeleteNote(noteId, email, cancellationToken);
+
+            return success
+                ? new ResultHandler(HttpStatusCode.NoContent)
+                : new ResultHandler(HttpStatusCode.NotFound);
         }
     }
 }
